+++ src/redturtle/prenotazioni/content/prenotazioni_folder.py
@@ -12,12 +12,9 @@ from z3c.form import validator
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from zope import schema
 from zope.component import provideAdapter
-from zope.i18n import translate
 from zope.interface import Invalid
 from zope.interface import implementer
 from zope.interface import invariant
-from zope.interface import provider
-from zope.schema.interfaces import IContextAwareDefaultFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
 
@@ -128,84 +125,6 @@ class IPauseTableRow(model.Schema):
         title=_("pause_end_label", default="Pause end"),
         vocabulary="redturtle.prenotazioni.VocOreInizio",
         required=False,
-    )
-
-
-@provider(IContextAwareDefaultFactory)
-def notify_on_submit_subject_default_factory(context):
-    return getattr(context, "translate", translate)(
-        _("notify_on_submit_subject_default_value", "Booking created ${title}")
-    )
-
-
-@provider(IContextAwareDefaultFactory)
-def notify_on_submit_message_default_factory(context):
-    return getattr(context, "translate", translate)(
-        _(
-            "notify_on_submit_message_default_value",
-            "Booking ${booking_type} for ${booking_date} at ${booking_time} was created.<a href=${booking_print_url}>Link</a>",
-        )
-    )
-
-
-@provider(IContextAwareDefaultFactory)
-def notify_on_confirm_subject_default_factory(context):
-    return getattr(context, "translate", translate)(
-        _(
-            "notify_on_confirm_subject_default_value",
-            "Booking of ${booking_date} at ${booking_time} was accepted",
-        )
-    )
-
-
-@provider(IContextAwareDefaultFactory)
-def notify_on_confirm_message_default_factory(context):
-    return getattr(context, "translate", translate)(
-        _(
-            "notify_on_confirm_message_default_value",
-            "The booking${booking_type} for ${title} was confirmed! <a href=${booking_print_url}>Link</a>",
-        )
-    )
-
-
-@provider(IContextAwareDefaultFactory)
-def notify_on_move_subject_default_factory(context):
-    return getattr(context, "translate", translate)(
-        _(
-            "notify_on_move_subject_default_value",
-            "Modified the boolking date for ${title}",
-        )
-    )
-
-
-@provider(IContextAwareDefaultFactory)
-def notify_on_move_message_default_factory(context):
-    return getattr(context, "translate", translate)(
-        _(
-            "notify_on_move_message_default_value",
-            "The booking scheduling of ${booking_type} was modified."
-            "The new one is on ${booking_date} at ${booking_time}. <a href=${booking_print_url}>Link</a>.",
-        )
-    )
-
-
-@provider(IContextAwareDefaultFactory)
-def notify_on_refuse_subject_default_factory(context):
-    return getattr(context, "translate", translate)(
-        _(
-            "notify_on_refuse_subject_default_value",
-            "Booking refused for ${title}",
-        )
-    )
-
-
-@provider(IContextAwareDefaultFactory)
-def notify_on_refuse_message_default_factory(context):
-    return getattr(context, "translate", translate)(
-        _(
-            "notify_on_refuse_message_default_value",
-            "The booking ${booking_type} of ${booking_date} at ${booking_time} was refused.",
-        )
     )
 
 
@@ -521,79 +440,6 @@ class IPrenotazioniFolder(model.Schema):
         default=False,
         required=False,
     )
-    notify_on_submit_subject = schema.TextLine(
-        title=_(
-            "notify_on_submit_subject",
-            default="Prenotazione created notification subject.",
-        ),
-        description=_("notify_on_submit_subject_help", default=""),
-        defaultFactory=notify_on_submit_subject_default_factory,
-        required=False,
-    )
-    notify_on_submit_message = schema.Text(
-        title=_(
-            "notify_on_submit_message",
-            default="Prenotazione created notification message.",
-        ),
-        description=_("notify_on_submit_message_help", default=""),
-        defaultFactory=notify_on_submit_message_default_factory,
-        required=False,
-    )
-    notify_on_confirm_subject = schema.TextLine(
-        title=_(
-            "notify_on_confirm_subject",
-            default="Prenotazione confirmed notification subject.",
-        ),
-        description=_("notify_on_confirm_subject_help", default=""),
-        defaultFactory=notify_on_confirm_subject_default_factory,
-        required=False,
-    )
-    notify_on_confirm_message = schema.Text(
-        title=_(
-            "notify_on_confirm_message",
-            default="Prenotazione confirmed notification message.",
-        ),
-        description=_("notify_on_confirm_message_help", default=""),
-        defaultFactory=notify_on_confirm_message_default_factory,
-        required=False,
-    )
-    notify_on_move_subject = schema.TextLine(
-        title=_(
-            "notify_on_move_subject",
-            default="Prenotazione moved notification subject.",
-        ),
-        description=_("notify_on_move_subject_help", default=""),
-        defaultFactory=notify_on_move_subject_default_factory,
-        required=False,
-    )
-    notify_on_move_message = schema.Text(
-        title=_(
-            "notify_on_move_message",
-            default="Prenotazione moved notification message.",
-        ),
-        description=_("notify_on_move_message_help", default=""),
-        defaultFactory=notify_on_move_message_default_factory,
-        required=False,
-    )
-    notify_on_refuse_subject = schema.TextLine(
-        title=_(
-            "notify_on_refuse_subject",
-            default="Prenotazione refused notification subject.",
-        ),
-        description=_("notify_on_refuse_subject_help", default=""),
-        defaultFactory=notify_on_refuse_subject_default_factory,
-        required=False,
-    )
-    notify_on_refuse_message = schema.Text(
-        title=_(
-            "notify_on_refuse_message",
-            default="Prenotazione created notification message.",
-        ),
-        description=_("notify_on_refuse_message_help", default=""),
-        defaultFactory=notify_on_refuse_message_default_factory,
-        required=False,
-    )
-
     max_bookings_allowed = schema.Int(
         title=_(
             "max_bookings_allowed_label",
@@ -605,6 +451,18 @@ class IPrenotazioniFolder(model.Schema):
         ),
         required=False,
         default=0,
+    )
+    reminder_notification_gap = schema.Int(
+        title=_(
+            "reminder_notification_gap_label",
+            default="Upcoming bookings notification gap",
+        ),
+        description=_(
+            "reminder_notification_gap_description",
+            default="Indicates how many days before of a booking the user will be notified about it.",
+        ),
+        required=False,
+        default=3,
     )
 
     model.fieldset(
@@ -638,6 +496,14 @@ class IPrenotazioniFolder(model.Schema):
     )
 
     model.fieldset(
+        "Reminders",
+        label=_("reminders_label", default="Reminders"),
+        fields=[
+            "app_io_enabled",
+        ],
+    )
+
+    model.fieldset(
         "Notifications",
         label=_("notifications_label", default="Notifications"),
         fields=[
@@ -646,61 +512,9 @@ class IPrenotazioniFolder(model.Schema):
             "notify_on_move",
             "notify_on_refuse",
         ],
-    )
-    model.fieldset(
-        "Prenotazioni Email Templates",
-        label=_(
-            "prenotazioni_email_templates_label",
-            default="Testo delle email di notifica",
+        description=_(
+            "notifications_help", default="Notifications management controlpanel"
         ),
-        # TODO: Use custom frontend widget for the new
-        # field where we must render the html of field's description
-        # description=_(
-        #     "templates_usage_default_value",
-        #     "${title} - title."
-        #     "${booking_gate} - booking gate."
-        #     "${booking_human_readable_start} - booking human readable start."
-        #     "${booking_date} - booking date."
-        #     "${booking_end_date} - booking end date."
-        #     "${booking_time} - booking time."
-        #     "${booking_time_end} - booking time end."
-        #     "${booking_code} - booking code."
-        #     "${booking_type} - booking type."
-        #     "${booking_print_url} - booking print url."
-        #     "${booking_url_with_delete_token} - booking url with delete token."
-        #     "${booking_user_phone} - booking user phone."
-        #     "${booking_user_email} - booking user email."
-        #     "${booking_user_details} - booking user details."
-        #     "${booking_office_contact_phone} - booking office contact phone."
-        #     "${booking_office_contact_pec} - booking office contact pec."
-        #     "${booking_office_contact_fax} - booking office contact fax."
-        #     "${booking_how_to_get_to_office} - booking how to get to office."
-        #     "${booking_office_complete_address} - booking office complete address.",
-        #     "${booking_user_details} - booking user details",
-        #     "${booking_requirements} - booking requeirements.",
-        #     "${prenotazioni_folder_title} - prenotazioni folder title.",
-        # ),
-        # description=_(
-        #     "prenotazioni_email_templates_description",
-        #     default="",
-        # ),
-        fields=[
-            "notify_on_submit_subject",
-            "notify_on_submit_message",
-            "notify_on_confirm_subject",
-            "notify_on_confirm_message",
-            "notify_on_move_subject",
-            "notify_on_move_message",
-            "notify_on_refuse_subject",
-            "notify_on_refuse_message",
-        ],
-    )
-    model.fieldset(
-        "Reminders",
-        label=_("reminders_label", default="Reminders"),
-        fields=[
-            "app_io_enabled",
-        ],
     )