+++ src/redturtle/prenotazioni/scripts/app_io.py
@@ -12,9 +12,8 @@ from zope.annotation.interfaces import IAnnotations
 
 from redturtle.prenotazioni.config import NOTIFICATIONS_LOGS
 from redturtle.prenotazioni.config import VERIFIED_BOOKING
-
-from .io_tools.api import Api
-from .io_tools.storage import logstorage
+from redturtle.prenotazioni.io_tools.api import Api
+from redturtle.prenotazioni.io_tools.storage import logstorage
 
 logger = logging.getLogger("redturtle.prenotazioni.app_io")
 locale.setlocale(locale.LC_ALL, "it_IT.UTF-8")