+++ src/redturtle/prenotazioni/vocabularies/tipologies.py
@@ -25,13 +25,12 @@ class PrenotazioneTypesVocabulary(object):
         """The vocabulary terms"""
         prenotazioni_folder = getPrenotazioniFolder(context)
 
-        res = [
+        return [
             self.booking_type2term(booking_type)
             for booking_type in prenotazioni_folder
             and prenotazioni_folder.get_booking_types()
             or []
         ]
-        return res
 
     def __call__(self, context):
         """