+++ CHANGES.rst
@@ -5,7 +5,11 @@ Changelog
 2.2.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Split booking notification gateways to 'App IO', 'Email' and 'SMS'
+  [folix-01]
+
+- Add booking reminder emails.
+  [folix-01]
 
 
 2.2.5 (2023-12-04)
@@ -40,6 +44,7 @@ Changelog
 
 - Set PrenotazioniFolder and PrenotazioneType as not searchable (types_not_searched).
   [cekk]
+
 - Send ical also for manager notifications.
   [cekk]
 
@@ -77,7 +82,7 @@ Changelog
 
 - Add booking_refuse_message to Prenotazione stringinterp variables.
   [folix-01]
- 
+
 - Extended PrenotaizoniFolder email templates var list.
   [folix-01]